added settings and separated std plugins
