+++ honeybot/settings/CONNECT.conf
