+++ honeybot/settings/STD_PLUGINS.conf
